+++ documentation/ChangeLog.md
@@ -1,6 +1,9 @@
 ## LANG_IDE_NAME release ChangeLog
+[Latest features on top]
 
 ### LANG_IDE_NAME 0.0.x
+ * Added editor Go To Matching Bracket action (shortcut: Ctrl+Shift+P)
+
  * Added New Project wizard.
  
   ▶ Recommended/tested CDT version: 8.5.0